requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
defusedxml==0.7.1
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-xml==2.0.0
idna==3.4
iniconfig==1.1.1
numpy==1.23.3
packaging==21.3
pandas==1.5.0
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.3
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
PyVCF3==1.0.3
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.12
